+++ testing/hometown-hero-webapp/src/pages/VNodeHome.tsx
@@ -1,5 +1,6 @@
 import Chat from '@/components/VNodeHome/Chat/Chat';
 import ChatParticipants from '@/components/VNodeHome/Chat/ChatParticipants';
+import ChatWrapper from '@/components/VNodeHome/Chat/ChatWrapper';
 import NodeDetails from '@/components/VNodeHome/NodeDetails/NodeDetails';
 import PageContentWrapper from '@/components/PageContentWrapper/PageContentWrapper';
 import PageHeader from '@/components/PageHeader/PageHeader';
@@ -9,35 +10,11 @@ import VisualizationWrapper from '@/components/Visualizations/VisualizationWrapp
 import { useState } from 'react';
 
 const VNodeHome = () => {
-    const [selectedParticipants, setSelectedParticipants] = useState<string[]>([]);
-
-    const handleSelectReplyParticipant = (participant: string) => {
-        setSelectedParticipants([participant]);
-    };
-
-    console.log('VNODE HOME', selectedParticipants);
-
     return (
         <PageContentWrapper>
             <div className="mt-24 sm:mt-0 md:mt-0 lg:mt-0" />
             <PageHeader withBackButton>V-Node Home</PageHeader>
-            <div className="flex flew-col gap-4 mt-8 ml-4 flex-wrap">
-                <Section title={'Chat'}>
-                    <Chat
-                        selectedParticipants={selectedParticipants}
-                        handleSelectReplyParticipant={handleSelectReplyParticipant}
-                    />
-                </Section>
-                <Section title={'Participants'}>
-                    <ChatParticipants
-                        selectedParticipants={selectedParticipants}
-                        setSelectedParticipants={setSelectedParticipants}
-                    />
-                </Section>
-                <Section title={'VNode Details'}>
-                    <NodeDetails />
-                </Section>
-            </div>
+            <ChatWrapper />
             <VisualizationWrapper width={520}>
                 <VNodeHomeViz />
             </VisualizationWrapper>
